# public/openapi.yaml
openapi: 3.0.3
info:
  title: Fine-Tune Labs API
  version: 1.0.0
  description: |

    # Fine-Tune Labs Public API

    Production-grade API for LLMOps, model training, and monitoring.

    ## Features

    - 🎯 **Widget Apps** - Deploy production monitoring widgets
    - 📊 **Analytics & Metrics** - Real-time model performance tracking
    - 🚀 **Inference Deployment** - Deploy and manage model endpoints
    - 🤖 **Model Management** - CRUD operations for LLM models
    - 📈 **Training Jobs** - Monitor fine-tuning progress

    ## Authentication

    All endpoints require Bearer token authentication:

    ```
    Authorization: Bearer YOUR_API_TOKEN
    ```

    Get your API token from the Fine-Tune Labs dashboard.

    ## Base URL

    - **Development**: `http://localhost:3000`
    - **Production**: `https://finetunelab.ai`

    ## Rate Limits

    - 1000 requests per minute per API key
    - 10,000 requests per day per API key

    ## Support

    - Documentation: https://finetunelab.ai/docs
    - Email: support@finetunelab.ai
  contact:
    name: Fine-Tune Labs API Support
    email: support@finetunelab.ai
    url: https://finetunelab.ai/support
  license:
    name: Proprietary
    url: https://finetunelab.ai/terms
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://finetunelab.ai
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Supabase authentication
...
security:
  - bearerAuth: []
...
paths:
...
  /api/inference/deploy:
    post:
      summary: Deploy model to production inference endpoint
      description: |
        Deploy your trained models to serverless GPU inference endpoints.

        This endpoint creates a production-ready, auto-scaling inference deployment on RunPod Serverless.
        Your model will be deployed with:
        - Auto-scaling GPU workers (scale to zero when idle)
        - Budget controls and spending limits
        - Production-grade load balancing
        - Per-request billing

        **Use Cases:**
        - Deploy fine-tuned models to production
        - Serve custom LLMs at scale
        - A/B test different model versions
        - Control infrastructure costs with budget limits

        **Workflow:**
        1. Train your model or upload model artifacts
        2. Store model weights in cloud storage (S3, GCS, etc.)
        3. Call this endpoint with model URL and config
        4. Use the returned endpoint_url for inference requests

        **Note:** Requires RunPod API key configured in Settings > Secrets
      tags:
        - Inference
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - provider
                - deployment_name
                - base_model
                - model_type
                - model_storage_url
                - budget_limit
              properties:
                provider:
                  type: string
                  enum:
                    - runpod-serverless
                  description: Infrastructure provider
                  example: runpod-serverless
                deployment_name:
                  type: string
                  description: Unique name for this deployment
                  example: my-finetuned-llama-v1
                base_model:
                  type: string
                  description: Base model architecture
                  example: meta-llama/Llama-2-7b-hf
                model_type:
                  type: string
                  enum:
                    - base
                    - finetuned
                    - quantized
                  description: Type of model deployment
                  example: finetuned
                model_storage_url:
                  type: string
                  format: uri
                  description: Cloud storage URL where model weights are stored (S3, GCS, etc.)
                  example: s3://my-bucket/models/llama2-finetuned/
                training_config_id:
                  type: string
                  description: Associated training configuration ID (optional)
                  example: config_abc123
                training_job_id:
                  type: string
                  description: Associated training job ID (optional)
                  example: job_xyz789
                model_artifact_id:
                  type: string
                  description: Model artifact ID from training (optional)
                  example: artifact_def456
                gpu_type:
                  type: string
                  enum:
                    - NVIDIA_A100_80GB
                    - NVIDIA_A40
                    - NVIDIA_RTX_A6000
                    - NVIDIA_RTX_A5000
                  description: GPU type for inference workers (defaults to optimal for model size)
                  example: NVIDIA_A100_80GB
                min_workers:
                  type: integer
                  default: 0
                  description: Minimum number of workers (0 = scale to zero when idle)
                  example: 0
                max_workers:
                  type: integer
                  default: 3
                  description: Maximum number of workers for auto-scaling
                  example: 5
                budget_limit:
                  type: number
                  description: Monthly budget limit in USD
                  example: 100
                auto_stop_on_budget:
                  type: boolean
                  default: true
                  description: Automatically stop deployment when budget is reached
                  example: true
                environment_variables:
                  type: object
                  description: Custom environment variables for inference container
                  additionalProperties:
                    type: string
                  example:
                    MAX_BATCH_SIZE: '8'
                    TEMPERATURE: '0.7'
      responses:
        '200':
          description: Deployment created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  deployment_id:
                    type: string
                    description: Unique deployment identifier
                    example: dep_abc123xyz
                  endpoint_url:
                    type: string
                    format: uri
                    description: Production inference endpoint URL
                    example: https://api.runpod.ai/v2/my-endpoint-abc123/runsync
                  status:
                    type: string
                    enum:
                      - INITIALIZING
                      - RUNNING
                      - STOPPED
                      - FAILED
                    description: Current deployment status
                    example: INITIALIZING
        '400':
          description: Bad request - Missing required fields or invalid provider
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        enum:
                          - VALIDATION_ERROR
                          - UNSUPPORTED_PROVIDER
                        example: VALIDATION_ERROR
                      message:
                        type: string
                        example: 'Missing required field: budget_limit'
        '401':
          description: Unauthorized - Missing or invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        enum:
                          - UNAUTHORIZED
                          - NO_CREDENTIALS
                        example: NO_CREDENTIALS
                      message:
                        type: string
                        example: No RunPod API key found
                      details:
                        type: string
                        example: Please add your RunPod API key in Settings > Secrets
        '500':
          description: Internal server error - Deployment failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        enum:
                          - DEPLOYMENT_FAILED
                          - INTERNAL_ERROR
                        example: DEPLOYMENT_FAILED
                      message:
                        type: string
                        example: Failed to create inference endpoint
                      details:
                        type: string
                        example: 'RunPod API error: Invalid GPU type'
...